requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
djangorestframework==3.14.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
Werkzeug==2.2.2
